Update  ASVP-ESD.v1 Parallel CNNs with Transformer Encoder and Co-Attention
This commit introduces a Python script that extracts features from audio files for speech emotion recognition tasks. 
The script includes functions to compute MFCC and Mel spectrograms using the `librosa` library. 
The output features are combined and saved for later use in model training.  
We excluded the disgust utterances in the classification experiment because there was a small number of them.